requirements.txt
tensorflow==2.16.1
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
